Readd CLONE_ABORT and add clone API
The reason readding the CLONE_CANCELLED to CLONE_ABORT:
Need check the target clone state in API, so the CLONE_SUCCESS
and CLONE_ABORT should be as the target clone states pool.
Rename to CLONE_ABORT to coresspond to the abort operation in clone.

Also add the new API, still need further design.